+++ CHANGELOG.md
@@ -6,17 +6,31 @@
 ### BREAKING CHANGES
 -->
 
-<a name="1.0.1"></a>
+<a name="1.0.2"></a>
+
+## 1.0.1 (2017-11-27)
+
+<a name="1.0.2"></a>
+
+### Features
+
+ * @AuditSynchronous interceptor to prevent asynchronous processing of audit events to avoid null-references in huge transactions.
+ * Interceptors are now based on JTA transactions
 
 ### Bug Fixes
 
- * Null check at @AuditRelatedEntity
+ * Fixed getRelatedEntityValue() & toAuditableString()
+
+
+<a name="1.0.1"></a>
  
 ## 1.0.1 (2017-11-17)
 
 ### Bug Fixes
 
-Initial release:
+* Null check at @AuditRelatedEntity
+
+
 
 <a name="1.0.0"></a>